+++ packages/basic-dialog/src/styles.js
@@ -1,23 +1,25 @@
 import { makeStyles } from '@material-ui/core/styles';
 
-const containerStyles = {
+const containerStyles = (theme) => ({
+    backgroundColor: theme.palette.background.default,
     borderRadius: 20,
     padding: '0 5%',
-};
+});
 
 const useStyles = makeStyles((theme) => ({
     contentContainer: {
         padding: 0,
     },
     largeContainer: {
-        ...containerStyles,
+        ...containerStyles(theme),
         minWidth: '60%',
     },
     mobileContainer: {
+        backgroundColor: theme.palette.background.default,
         padding: '0 5%',
     },
     smallContainer: {
-        ...containerStyles,
+        ...containerStyles(theme),
         minWidth: 600,
     },
     titleContainer: {